+++ src/main/java/com/nowui/chuangshi/service/BillCommissionService.java
@@ -17,7 +17,7 @@ public class BillCommissionService extends Service {
         return billCommissionCache.listByOrBill_idOrMember_id(bill_id, member_id);
     }
 
-    public Boolean save(String bill_id, String product_sku_id, String member_id, String member_name, String member_level_id, String member_level_name, Integer product_sku_commission,
+    public Boolean save(String bill_id, String product_sku_id, String member_id, String member_name, String member_level_id, String member_level_name, BigDecimal product_sku_commission,
             BigDecimal product_sku_commission_amount, String system_create_user_id) {
         return billCommissionCache.save(bill_id, product_sku_id, member_id, member_name, member_level_id, member_level_name, product_sku_commission, product_sku_commission_amount,
                 system_create_user_id);
